Set Proper Location
